requirements.txt
# LunarTech AI - RAG Chat & Handbook Generator
# Core
streamlit>=1.40.0
python-dotenv>=1.0.0

# PDF Processing
pdfplumber>=0.11.0

# Document Formats
python-docx>=1.1.0

# RAG Engine
lightrag-hku>=1.0.0

# LLM & Embeddings (OpenRouter via OpenAI SDK)
openai>=1.0.0
tiktoken>=0.7.0

# Database
supabase>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Utilities
numpy>=1.26.0

# OS & Voice Integration
psutil>=5.9.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.14

# UI Enhancements
streamlit-echarts>=0.5.0
websockets>=12.0
